Finished for LoginPage.jsx and CreateAccountPage.jsx
