+++ src/Components/CreateAccountPage.jsx
@@ -1,48 +1,8 @@
-import React, { useState } from 'react';
 
-function CreateAccountPage({ loginHandle, users, setUsers }) {
+function CreateAccountPage({ pageHandler, createAccountHandler, setUsername, setPassword, error }) {
 
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
+  //NO LOGIC HERE, KEEP CLEAN
 
-    //handle account creation
-
-   //handle account creation
-  const handleCreateAccount = () => {
-    // Basic validation
-    if (!username || !password) {
-      setError('Please fill in all fields.');
-      return;
-    }
-
-    // Ensure Password is more than 8 characters
-    if (password.length < 8) {
-        setError('Password must be at least 8 characters');
-        return;
-    }
-
-    // Check if the username already exists in the Map
-    if (users.has(username)) {
-      setError('Username already exists. Please choose a different one.');
-      return;
-    }
-
-    // Store the new username and password in the Map
-    setUsers((prevUsers) => {
-      const updatedUsers = new Map(prevUsers);
-      updatedUsers.set(username, password);
-      return updatedUsers;
-    });
-
-    // Reset fields and navigate to the login page
-    setUsername('');
-    setPassword('');
-    setError('');
-    loginHandle('login');
-  };
-
-    //various details
   return (
     <div style={{
       display: 'flex',
@@ -67,14 +27,12 @@ function CreateAccountPage({ loginHandle, users, setUsers }) {
           </div>
           <div>
             <input
-              value={username}
               placeholder="Enter username"
               onChange={(ev) => setUsername(ev.target.value)}
             />
           </div>
           <div>
             <input
-              value={password}
               type="password"
               placeholder="Enter password here"
               onChange={(ev) => setPassword(ev.target.value)}
@@ -84,7 +42,7 @@ function CreateAccountPage({ loginHandle, users, setUsers }) {
             <input
               type="button"
               value="Create Account"
-              onClick={handleCreateAccount}
+              onClick={createAccountHandler}
             />
           </div>
           {error && (
@@ -99,7 +57,7 @@ function CreateAccountPage({ loginHandle, users, setUsers }) {
             <input
               type="button"
               value="Log in"
-              onClick={() => loginHandle('login')}
+              onClick={() => pageHandler('login')}
             />
           </div>
         </div>
